Improve DefaultModuleIdentifier equals and hashCode
Equals now compares the name before the group, as there is less chance
of overlap for artifact names than group.
The hashCode is now cached at construction since the object is placed
in datastructure requiring it a lot.

Based on this change, added @Nullable on ModuleIdentifier.getGroup and
added an assertion to the implementation.

Some of these performance changes make up for the extra work
performed on selector removal.