Support per-task timeouts
Users can now configure a timeout on a per-task level.
If the task exceeds this timeout while running, its thread
is interrupted and it is marked as failed. This works for any
task that handles InterruptedExceptions correctly, including
tasks using our exec() infrastructure or worker API.

A timed-out task behaves just like a failed task in all other
respects, e.g. finalizers will still be run. If --continue is
used, other tasks can continue running after it.

Tasks that don't respond to interrupts are not yet handled.
Those should result in a shutdown of the daemon, as there is no
other way to restore a working system.